MeeseOS localStorage prefix
This is a better alternative to the previous implementation that was a workaround specifically for the emulator cookies, and it will allow for the further addition of third party data in localStorage without any parsing warnings.

+++ backend/server/src/auth.js
@@ -213,8 +213,8 @@ class Auth {
 
 	/**
 	 * Creates user profile object
-	 * @param {object} fields Input fields
-	 * @param {object} result Login result
+	 * @param {Object} fields Input fields
+	 * @param {Object} result Login result
 	 * @return {AuthUserProfile|boolean}
 	 */
 	createUserProfile(fields, result) {
@@ -274,9 +274,11 @@ class Auth {
 			try {
 				const { path, contents = "" } = file;
 				const shortcutsFile = await vfs.realpath(`home:/${path}`, profile);
-
-				await fs.ensureDir(pathLib.dirname(shortcutsFile));
-				await fs.writeFile(shortcutsFile, contents);
+				const dir = pathLib.dirname(shortcutsFile);
+				if (!await fs.pathExists(shortcutsFile)) {
+					await fs.ensureDir(dir);
+					await fs.writeFile(shortcutsFile, contents);
+				}
 			} catch (e) {
 				console.warn(`There was a problem writing '${file.path}' to the home directory template`);
 				console.error("ERROR:", e);
